big refactor migrates swerve mpc into a class
